removed throw specification for unit_test::do_test
I have always disliked the throw specification for unit_test::do_test,
so I have now removed it.
An added benefit is that unit tests can now throw whatever,
and the test_suite will now catch it.
This means that if tested code throws and the exception is not picked up by
unit test, the unit test will fail with an appropriate output, but the
test suite will continue. Awesome!